Chore: Scout 페이지 스타일 코드 분리
